Revert "fix(mfa): wait for account data before deciding on registration"
This reverts commit cfc1273501c7b379d1eacf6a2ae76f964029ff38. The gate
can wait without bound when OpenApp never completes (e.g. offline), so
the fix will return in a follow-up together with a bounded wait.
