+++ src/components/MultifactorAuthentication/biometrics/operations/index.native.ts
@@ -1,8 +1,7 @@
 import addMFABreadcrumb from '@components/MultifactorAuthentication/observability/breadcrumbs';
 
 import {decodeLibraryError, getKeyAlias} from '@libs/MultifactorAuthentication/NativeBiometricsHSM/helpers';
-import waitForAccountDataReady from '@libs/MultifactorAuthentication/shared/waitForAccountDataReady';
-import {readOnyxValueOnce} from '@libs/MultifactorAuthentication/shared/waitForOnyxValue';
+import readOnyxValueOnce from '@libs/MultifactorAuthentication/shared/readOnyxValueOnce';
 
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
@@ -55,7 +54,6 @@ async function areLocalCredentialsKnownToServer(accountID: number, signal?: Abor
     if (!localCredentialID) {
         return false;
     }
-    await waitForAccountDataReady(signal);
     const account = await readOnyxValueOnce(ONYXKEYS.ACCOUNT, signal);
     return (mfaCredentialIDsSelector(account) ?? []).includes(localCredentialID);
 }
